Add libcrypto for users of libstagefright.
libstagefright_rtsp uses some MD5 functions that used to be in bionic,
but it was removed recently. As an initial fix, I statically linked in
libcrypto_static to the libstagefright_rtsp library. However, I think
it's better to modify the single user of this library to link against
the shared libcrypto library.

Change-Id: Iaf2e1aeea32fd8af038f6e77bf58ea7df50d807a